One function answers "does this project use remote files"
There were two, and they disagreed about `s.images({ remote: true })`.

`hasRemoteFileSchema` (server, gating whether `/save` demands remote credentials)
only recursed into a record's `item`. A media collection serializes as a `record`
whose item is an object of width/height/mimeType/alt, with the file named by the
record's KEY — so there is no image schema inside to find, and it returned false.
`findRequiredRemoteFiles` (Studio, gating the `/remote/settings` fetch) had a
`mediaType && remote` branch and returned true for the same schema.

The server's answer was the wrong one, and it fails silently. With it false,
`/save` runs `saveOrUploadFiles` in `skip-remote` mode — and that mode does not
merely skip the upload: the loop over remote descriptors sits inside the
`upload-remote` branch, so every remote file is dropped with no error. The commit
then lands a remote ref with no bytes behind it: a broken image, committed to the
repository, with nothing anywhere saying why.

One implementation now, in `@valbuild/core`, where both realms can reach it —
both callers hold a `SerializedSchema` rather than a `Schema`, so this stays a
walk over the serialized form. `@valbuild/server` still exports
`hasRemoteFileSchema` (now a re-export), so no import changes for consumers.

Two behaviours had to be chosen rather than merged:

- **The record branch** takes the Studio's `mediaType && remote` shortcut AND
  still recurses into `item`, so an ordinary record holding a remote field is
  found too. The old Studio version returned early only on the media case, which
  is right; the old server version only recursed, which is what missed it.
- **An unknown schema type throws**, as the server's did, rather than logging and
  returning null as the Studio's did. Returning false is the silent skip above, so
  it is not a safe default. The `never` assignment makes forgetting a new schema
  type a compile error; the throw is for data that should be impossible. The
  Studio catches at its one call site, where the cost of throwing is the whole UI
  and all that is at stake is whether to fetch remote settings.

The tests are merged into one file in core and the media-collection cases are new
— neither old test had any. They build fixtures with `s` and serialize them, which
is the point: the bug was a walk that did not match the shape `s.images()` really
produces, and a hand-written fixture would have agreed with the walk instead of
with the schema. Verified against the old implementation first: it returns false
for both `s.images({ remote: true })` and a remote `s.files()`.
